EC Write Added

+++ src/edu/rutgers/winlab/sim/core/ISerializable.java
@@ -2,6 +2,6 @@ package edu.rutgers.winlab.sim.core;
 
 public interface ISerializable {
 	
-	public int getSizeInBits();
+	public long getSizeInBits();
 
 }
